pyproject.toml
[tool.poetry]
name = "jobboard"
version = "0.1.0"
description = ""
authors = ["Kerelko Barabash <kerelkobarabash@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-ninja = "^1.1.0"
factory-boy = "^3.3.0"
pytest = "^8.1.1"
pytest-django = "^4.8.0"
punq = "^0.7.0"
celery = "^5.4.0"
sqlalchemy = "^2.0.29"
psycopg2-binary = "^2.9.9"
httpx = "^0.27.0"
pre-commit = "^3.7.0"
ruff = "^0.4.3"
redis = "^5.0.4"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
DJANGO_SETTINGS_MODULE = "src.core.settings.prod"

[tool.ruff]
line-length = 82
indent-width = 4
show-fixes = true

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"foofile.py" = ["E701"]
